requirements.txt
httpx>=0.27
openai>=1.40
python-dotenv>=1.0
